Updates in EntryStore
- update p2panda-rs
- restructure modules
- introduce new replication methods
