+++ aquadoggo/src/db/models/mod.rs
@@ -1,7 +1,4 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-mod entry;
-mod log;
-
-pub use self::log::Log;
-pub use entry::EntryRow;
+pub mod entry;
+pub mod log;
