+++ aquadoggo/src/db/stores/mod.rs
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+mod entry;
+mod log;
+#[cfg(test)]
+mod test_utils;
