+++ aquadoggo/src/rpc/methods/entry_args.rs
@@ -3,7 +3,7 @@
 use jsonrpc_v2::{Data, Params};
 use p2panda_rs::storage_provider::traits::StorageProvider;
 
-use crate::db::store::SqlStorage;
+use crate::db::provider::SqlStorage;
 use crate::errors::StorageProviderResult;
 use crate::rpc::request::EntryArgsRequest;
 use crate::rpc::response::EntryArgsResponse;
@@ -76,8 +76,8 @@ mod tests {
             r#"{
                 "entryHashBacklink": null,
                 "entryHashSkiplink": null,
-                "seqNum": "1",
-                "logId": "1"
+                "seqNum": 1,
+                "logId": 1
             }"#,
         );
